pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rabbitmq-mcp-server"
version = "1.0.0"
description = "RabbitMQ Debug & Visualization MCP Server"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "tenacity>=8.2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
rabbitmq-mcp-server = "rabbitmq_mcp_server.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/rabbitmq_mcp_server"]
